requirements.txt
google-auth-oauthlib
google-api-python-client
rich
pathlib